testing application stop with if else

+++ scripts/application_stop.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 #Stopping existing node servers
 echo "Stopping any existing node servers"
-pkill node
+if pgrep -x "node" > /dev/null
+then
+    pkill node
+else
+    echo "already stopped"
+fi
